Apparent bug in want comments when used in a block comment
Currently fails with this additional error:
```diff
  a/b.go:7: in 'want' comment: got EOF, want regular expression
  a/b.go:8: in 'want' comment: invalid char escape
+ a/b.go:40: in 'want' comment: got Ident after odd, want ':'
  a/b.go:11:9: diagnostic "call of println(...)" does not match pattern `wrong expectation text`
  a/b.go:14:9: unexpected diagnostic: call of println(...)
```

I'm not really sure what's going on there, but in other tests I've made
I can see that a test file like this:
```go
package whatever

/* // want "one"
block comment // want "two"
*/
func x(){
        /*
                more comments // want "three"
        */
}
```
I get a few kinds of behavior with unclear causes.

In a "fuller" test I get a variety of results including:
- "unexpected diagnostic: two" on the line where `// want "two"` exists
- "diagnostic "two" does not match pattern "two" where `// want "two"` exists
- "no diagnostic was reported matching "two"" where `// want "two"` exists
- "in 'want' comment: got Ident after odd, want ':'" for the `// want "one"` line

That last one can be reproduced very easily: https://github.com/Groxx/analysisbug

I can attempt to find a minimal repro for the other three, but so far no luck
and there's a chance it'll be fixed in the process of finding the cause of the
"got Ident" issue, which seems like a pretty clear mistake.